changing submission.py removing the models and adding them to a deticated models.py file
